requirements.txt
djangorestframework==3.11.0
ipython==7.15.0
pytz==2020.1
requests==2.24.0
django-cors-headers==3.4.0
gunicorn==20.0.4
django-jsoneditor==0.1.6
psycopg2-binary==2.8.5
oauth2client==4.1.3
httplib2==0.18.1
Django==3.1.6
psycopg2-binary==2.8.5
django-redis==4.12.1
Pillow==8.1.0
